airflow/requirements.txt
requests
scrapy
ipython
python-dotenv
pandas
pyarrow

scrapy-deltafetch
google-cloud-bigquery
google-cloud-storage
apache-airflow-providers-google